# services/api/openapi.json
{
  "components": {
    "schemas": {
      "EstimateInterval": {
        "additionalProperties": false,
        "properties": {
          "calibration_method": {
            "title": "Calibration Method",
            "type": "string"
          },
          "confidence_level": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Confidence Level",
            "type": "number"
          },
          "p10": {
            "minimum": 0.0,
            "title": "P10",
            "type": "number"
          },
          "p50": {
            "minimum": 0.0,
            "title": "P50",
            "type": "number"
          },
          "p90": {
            "minimum": 0.0,
            "title": "P90",
            "type": "number"
          }
        },
        "required": [
          "p10",
          "p50",
          "p90",
          "confidence_level",
          "calibration_method"
        ],
        "title": "EstimateInterval",
        "type": "object"
      },
      "EstimateRequest": {
        "additionalProperties": false,
        "properties": {
          "matter_input": {
            "$ref": "#/components/schemas/MatterInput"
          },
          "model_strategy": {
            "enum": [
              "firm_specific",
              "pooled_research",
              "synthetic_baseline"
            ],
            "title": "Model Strategy",
            "type": "string"
          },
          "risk_tolerance": {
            "enum": [
              "conservative",
              "balanced",
              "aggressive"
            ],
            "title": "Risk Tolerance",
            "type": "string"
          },
          "tenant_id": {
            "minLength": 1,
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "matter_input",
          "risk_tolerance",
          "model_strategy"
        ],
        "title": "EstimateRequest",
        "type": "object"
      },
      "FeeRecommendation": {
        "additionalProperties": false,
        "properties": {
          "billing_model": {
            "enum": [
              "Hourly",
              "Fixed Fee",
              "Capped Fee",
              "Retainer",
              "Outcome Related"
            ],
            "title": "Billing Model",
            "type": "string"
          },
          "cap_amount_hkd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cap Amount Hkd"
          },
          "confidence_interval_high_hkd": {
            "exclusiveMinimum": 0.0,
            "title": "Confidence Interval High Hkd",
            "type": "number"
          },
          "confidence_interval_low_hkd": {
            "exclusiveMinimum": 0.0,
            "title": "Confidence Interval Low Hkd",
            "type": "number"
          },
          "downside_risk_hkd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downside Risk Hkd"
          },
          "expected_downside_hkd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Downside Hkd"
          },
          "expected_margin_hkd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Margin Hkd"
          },
          "margin_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin Pct"
          },
          "partner_decision_support_disclaimer": {
            "default": "This recommendation is decision-support only; a responsible partner must review matter context, professional obligations, and client instructions before use.",
            "title": "Partner Decision Support Disclaimer",
            "type": "string"
          },
          "pricing_guardrails": {
            "items": {
              "type": "string"
            },
            "title": "Pricing Guardrails",
            "type": "array"
          },
          "recommended_fee_hkd": {
            "exclusiveMinimum": 0.0,
            "title": "Recommended Fee Hkd",
            "type": "number"
          },
          "schema_version": {
            "const": "proforma.matter.v1",
            "default": "proforma.matter.v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "billing_model",
          "recommended_fee_hkd",
          "confidence_interval_low_hkd",
          "confidence_interval_high_hkd"
        ],
        "title": "FeeRecommendation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MatterInput": {
        "additionalProperties": false,
        "properties": {
          "billing_model": {
            "enum": [
              "Hourly",
              "Fixed Fee",
              "Capped Fee",
              "Retainer",
              "Outcome Related"
            ],
            "title": "Billing Model",
            "type": "string"
          },
          "client_type": {
            "enum": [
              "Mainland Enterprise",
              "HK Listed Co.",
              "SME/Local Business",
              "Individual",
              "Financial Institution",
              "SOE"
            ],
            "title": "Client Type",
            "type": "string"
          },
          "complexity_score": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Complexity Score",
            "type": "integer"
          },
          "cross_border_flag": {
            "title": "Cross Border Flag",
            "type": "boolean"
          },
          "deal_value_hkd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Value Hkd"
          },
          "document_volume": {
            "exclusiveMinimum": 0.0,
            "title": "Document Volume",
            "type": "integer"
          },
          "firm_tier": {
            "enum": [
              "Magic Circle / International",
              "PRC Elite Firm in HK",
              "Large Local (11+ partners)",
              "Mid-tier (6-10 partners)",
              "Small/Boutique (1-5 partners)"
            ],
            "title": "Firm Tier",
            "type": "string"
          },
          "jurisdiction": {
            "enum": [
              "HK Only",
              "GBA Cross-Border (HK-PRC)",
              "Multi-Jurisdictional (APAC)"
            ],
            "title": "Jurisdiction",
            "type": "string"
          },
          "matter_subtype": {
            "title": "Matter Subtype",
            "type": "string"
          },
          "matter_type": {
            "enum": [
              "M&A",
              "Litigation",
              "Commercial Property",
              "Employment",
              "IP/Technology",
              "Corporate Restructuring",
              "Banking & Finance",
              "Wills & Probate",
              "Regulatory/Compliance",
              "Arbitration"
            ],
            "title": "Matter Type",
            "type": "string"
          },
          "party_count": {
            "exclusiveMinimum": 0.0,
            "title": "Party Count",
            "type": "integer"
          },
          "risk_tolerance": {
            "enum": [
              "Low",
              "Medium",
              "High"
            ],
            "title": "Risk Tolerance",
            "type": "string"
          },
          "schema_version": {
            "const": "proforma.matter.v1",
            "default": "proforma.matter.v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "matter_type",
          "matter_subtype",
          "jurisdiction",
          "firm_tier",
          "client_type",
          "document_volume",
          "complexity_score",
          "party_count",
          "cross_border_flag",
          "billing_model",
          "risk_tolerance"
        ],
        "title": "MatterInput",
        "type": "object"
      },
      "PredictionResponse": {
        "additionalProperties": false,
        "properties": {
          "cost_estimate": {
            "$ref": "#/components/schemas/EstimateInterval"
          },
          "dataset_lineage": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Dataset Lineage",
            "type": "object"
          },
          "decision_support_disclaimer": {
            "title": "Decision Support Disclaimer",
            "type": "string"
          },
          "duration_estimate": {
            "$ref": "#/components/schemas/EstimateInterval"
          },
          "estimate_id": {
            "title": "Estimate Id",
            "type": "string"
          },
          "fee_recommendation": {
            "$ref": "#/components/schemas/FeeRecommendation"
          },
          "input_summary": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Input Summary",
            "type": "object"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "scope_creep_probability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Scope Creep Probability",
            "type": "number"
          },
          "stage_estimates": {
            "items": {
              "$ref": "#/components/schemas/StageEstimate"
            },
            "title": "Stage Estimates",
            "type": "array"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "estimate_id",
          "tenant_id",
          "model_version",
          "dataset_lineage",
          "input_summary",
          "cost_estimate",
          "duration_estimate",
          "scope_creep_probability",
          "stage_estimates",
          "fee_recommendation",
          "decision_support_disclaimer",
          "limitations"
        ],
        "title": "PredictionResponse",
        "type": "object"
      },
      "QuoteBenchmarkSegment": {
        "additionalProperties": false,
        "properties": {
          "dimensions": {
            "items": {
              "type": "string"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "fallback_level": {
            "title": "Fallback Level",
            "type": "string"
          },
          "sample_size": {
            "minimum": 0.0,
            "title": "Sample Size",
            "type": "integer"
          },
          "segment_label": {
            "title": "Segment Label",
            "type": "string"
          }
        },
        "required": [
          "segment_label",
          "dimensions",
          "sample_size",
          "fallback_level"
        ],
        "title": "QuoteBenchmarkSegment",
        "type": "object"
      },
      "QuotePackChartSpec": {
        "additionalProperties": false,
        "properties": {
          "chart_type": {
            "title": "Chart Type",
            "type": "string"
          },
          "data": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Data",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "chart_type",
          "title",
          "description",
          "data"
        ],
        "title": "QuotePackChartSpec",
        "type": "object"
      },
      "QuotePackMetric": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "display_value": {
            "title": "Display Value",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "sample_size": {
            "minimum": 0.0,
            "title": "Sample Size",
            "type": "integer"
          },
          "segment_label": {
            "title": "Segment Label",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "label",
          "value",
          "display_value",
          "description",
          "segment_label",
          "sample_size"
        ],
        "title": "QuotePackMetric",
        "type": "object"
      },
      "QuotePackRenderResponse": {
        "additionalProperties": false,
        "properties": {
          "approved_shareable": {
            "default": false,
            "title": "Approved Shareable",
            "type": "boolean"
          },
          "checksum_sha256": {
            "title": "Checksum Sha256",
            "type": "string"
          },
          "estimate_id": {
            "title": "Estimate Id",
            "type": "string"
          },
          "file_size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "File Size Bytes",
            "type": "integer"
          },
          "quote_pack_id": {
            "title": "Quote Pack Id",
            "type": "string"
          },
          "rendered_at": {
            "format": "date-time",
            "title": "Rendered At",
            "type": "string"
          },
          "snapshot_checksum": {
            "title": "Snapshot Checksum",
            "type": "string"
          },
          "status": {
            "const": "rendered",
            "title": "Status",
            "type": "string"
          },
          "storage_backend": {
            "enum": [
              "local",
              "supabase"
            ],
            "title": "Storage Backend",
            "type": "string"
          },
          "storage_path": {
            "title": "Storage Path",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "quote_pack_id",
          "estimate_id",
          "tenant_id",
          "status",
          "storage_backend",
          "storage_path",
          "checksum_sha256",
          "file_size_bytes",
          "snapshot_checksum"
        ],
        "title": "QuotePackRenderResponse",
        "type": "object"
      },
      "QuoteSubstantiationResponse": {
        "additionalProperties": false,
        "properties": {
          "assumptions_and_guardrails": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions And Guardrails",
            "type": "array"
          },
          "benchmark_segment": {
            "$ref": "#/components/schemas/QuoteBenchmarkSegment"
          },
          "chart_specs": {
            "items": {
              "$ref": "#/components/schemas/QuotePackChartSpec"
            },
            "title": "Chart Specs",
            "type": "array"
          },
          "estimate_id": {
            "title": "Estimate Id",
            "type": "string"
          },
          "evidence_footer": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Footer",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/QuotePackMetric"
            },
            "title": "Metrics",
            "type": "array"
          },
          "schema_version": {
            "const": "proforma.quote_pack.v1",
            "default": "proforma.quote_pack.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "snapshot_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Checksum"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "approved",
              "rendering",
              "rendered",
              "shared",
              "revoked",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "estimate_id",
          "tenant_id",
          "benchmark_segment",
          "metrics",
          "chart_specs",
          "assumptions_and_guardrails",
          "evidence_footer",
          "limitations"
        ],
        "title": "QuoteSubstantiationResponse",
        "type": "object"
      },
...
      "StageEstimate": {
        "additionalProperties": false,
        "properties": {
          "associate_hours": {
            "minimum": 0.0,
            "title": "Associate Hours",
            "type": "number"
          },
          "cost_hkd": {
            "minimum": 0.0,
            "title": "Cost Hkd",
            "type": "number"
          },
          "partner_hours": {
            "minimum": 0.0,
            "title": "Partner Hours",
            "type": "number"
          },
          "stage_name": {
            "title": "Stage Name",
            "type": "string"
          }
        },
        "required": [
          "stage_name",
          "partner_hours",
          "associate_hours",
          "cost_hkd"
        ],
        "title": "StageEstimate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ProForma API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/estimates": {
      "post": {
        "operationId": "create_estimate_v1_estimates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Estimate",
        "tags": [
          "estimates"
        ]
      }
    },
    "/v1/estimates/{estimate_id}": {
      "get": {
        "operationId": "get_estimate_v1_estimates__estimate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "estimate_id",
            "required": true,
            "schema": {
              "title": "Estimate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Estimate",
        "tags": [
          "estimates"
        ]
      }
    },
    "/v1/estimates/{estimate_id}/quote-packs/render": {
      "post": {
        "operationId": "render_quote_pack_v1_estimates__estimate_id__quote_packs_render_post",
        "parameters": [
          {
            "in": "path",
            "name": "estimate_id",
            "required": true,
            "schema": {
              "title": "Estimate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotePackRenderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render Quote Pack",
        "tags": [
          "estimates"
        ]
      }
    },
    "/v1/estimates/{estimate_id}/quote-substantiation": {
      "get": {
        "operationId": "get_quote_substantiation_v1_estimates__estimate_id__quote_substantiation_get",
        "parameters": [
          {
            "in": "path",
            "name": "estimate_id",
            "required": true,
            "schema": {
              "title": "Estimate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteSubstantiationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quote Substantiation",
        "tags": [
          "estimates"
        ]
      }
    },
    "/v1/estimates/{estimate_id}/quote-substantiation.pdf": {
      "get": {
        "operationId": "get_quote_substantiation_pdf_v1_estimates__estimate_id__quote_substantiation_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "estimate_id",
            "required": true,
            "schema": {
              "title": "Estimate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quote Substantiation Pdf",
        "tags": [
          "estimates"
        ]
      }
    },
...
  }
}